+++ tests/records_subscribe.rs
@@ -10,7 +10,7 @@ use serde_json::json;
 use vercre_dwn::data::DataStream;
 use vercre_dwn::provider::KeyStore;
 use vercre_dwn::records::{QueryBuilder, RecordsFilter, SubscribeBuilder, WriteBuilder, WriteData};
-use vercre_dwn::{endpoint, Message};
+use vercre_dwn::{Message, endpoint};
 
 // The owner should be able to to subscribe their own event stream
 #[tokio::test]
@@ -27,10 +27,7 @@ async fn owner_events() {
     let reply =
         endpoint::handle(ALICE_DID, subscribe, &provider).await.expect("should configure protocol");
     assert_eq!(reply.status.code, StatusCode::OK);
-
-    let Some(mut subscribe_reply) = reply.body else {
-        panic!("unexpected reply: {:?}", reply);
-    };
+    let mut subscribe_reply = reply.body.expect("should have body");
 
     // --------------------------------------------------
     // Alice writes a record.
